Extract Ingestor classes from Tio2Jira
This experiment effectively hollows out Tio2Jira and transforms it into
a factory for some Ingestor classes, which are specialised to TenableIO
and TenableSC.

I've left the `screen_builder()` as it felt more like a Jira
initialisation task than something belonging to an Ingestor.

We could have used the strategy pattern instead of inheritance, but this
seemed like the more elegant solution for a shallow, narrow set of
classes that won't be subclassed.

The __init__ arguments could do with some thought - I just added params
as they turned up.

NB. check the logic in the `create_issues()` and `close_issues()`
methods.