Jira’s API Docs are incorrect
This shoudl address #257, #258, and #259.  The issue seems to be related to the API docs referring to “issue id OR key” when in reality it only seems to update on the issue key.  Refactored the code to use the issue keys instead of the numerical IDs, and updating seems to be occuring as expected.

https://developer.atlassian.com/cloud/jira/platform/rest/v3/api-group-issues/#api-rest-api-3-issue-issueidorkey-put