Exception in subtask upsert was hidden by multi-threading #262
The SubTask upsert method was failing do to the mapping change from 2.0.4, however was masked because this was occuring within a worker thread.  Corrected the issue, Added logic to only use threading when the max_workers is above 1 (for testing), and also added a failsafe to stop execution after processing findings if the number of observed exceptions from the worker jobs is above 0.